+++ func/searchIFRA.php
@@ -1,7 +1,7 @@
 <?php 
 if (!defined('pvault_panel')){ die('Not Found');}
 
-function searchIFRA($cas, $name, $get, $conn, $cat = 'cat4'){
+function searchIFRA($cas, $name, $get, $conn, $defCatClass ){
 	if($cas !== '0'){//IGNORE VALUE FOR CARRIERS
 		if($cas){
 			$q = "cas = '$cas'";
@@ -9,15 +9,15 @@ function searchIFRA($cas, $name, $get, $conn, $cat = 'cat4'){
 			$q = "name = '$name' OR synonyms LIKE '%$name%'";
 		}
 			
-		$res = mysqli_fetch_array(mysqli_query($conn, "SELECT risk, $cat, type, formula FROM IFRALibrary WHERE $q"));
+		$res = mysqli_fetch_array(mysqli_query($conn, "SELECT risk, $defCatClass, type, formula FROM IFRALibrary WHERE $q"));
 		if($get){
 			return $res[$get];
 		}else{		
 			if($res){
-				if(!$res["$cat"]){
+				if(!$res["$defCatClass"]){
 					return $res['type'].' - '.$res['risk'];
 				}else{
-					return $res["$cat"].' - '.$res['risk'];
+					return $res["$defCatClass"].' - '.$res['risk'];
 				}
 			}
 		}