+++ pages/makeFormula.php
@@ -202,7 +202,7 @@ $(document).ready(function() {
                   <?php while ($formula = mysqli_fetch_array($formula_q)) {
 					 	$ing_q = mysqli_fetch_array(mysqli_query($conn, "SELECT cas, $defCatClass, price, ml, profile FROM ingredients WHERE BINARY name = '".$formula['ingredient']."'"));
 
-						$limitIFRA = searchIFRA($ing_q['cas'],$formula['ingredient'],null,$conn);
+						$limitIFRA = searchIFRA($ing_q['cas'],$formula['ingredient'],null,$conn,$defCatClass);
 						$limit = explode(' - ', $limitIFRA);
 					    $limit = $limit['0'];